split dashboard tests
