+++ tests/test_dashboard_setup.py
@@ -1,0 +1,84 @@
+"""Module to test the dashboard.
+
+This test module uses (playwright)[https://playwright.dev/python/]
+to test the user workflow.
+
+Installation:
+
+    pip install pytest-playwright
+    playwright install
+
+Make sure that the server is running by:
+```bash
+cd dianna/dashboard
+streamlit run Home.py
+```
+Then, set variable `LOCAL=True` (see below) to connect to local instance for
+debugging. Then, you can run the tests with:
+
+```bash
+pytest -v -m dashboard --dashboard
+```
+See more documentation about dashboard in: dianna/dashboard/readme.md
+
+For Code generation (https://playwright.dev/python/docs/codegen):
+
+    playwright codegen http://localhost:8501
+"""
+
+import time
+from contextlib import contextmanager
+import pytest
+from playwright.sync_api import Page
+from playwright.sync_api import expect
+
+LOCAL = False
+
+PORT = '8501' if LOCAL else '8502'
+BASE_URL = f'localhost:{PORT}'
+
+pytestmark = pytest.mark.dashboard
+
+
+@pytest.fixture(scope='module', autouse=True)
+def before_module():
+    """Run dashboard in module scope."""
+    with run_streamlit():
+        yield
+
+
+@contextmanager
+def run_streamlit():
+    """Run the dashboard."""
+    import subprocess
+
+    if not LOCAL:
+        p = subprocess.Popen([
+            'dianna-dashboard',
+            '--server.port',
+            PORT,
+            '--server.headless',
+            'true',
+        ])
+        time.sleep(5)
+
+    yield
+
+    if not LOCAL:
+        p.kill()
+
+
+def test_page_load(page: Page):
+    """Test performance of landing page."""
+    page.goto(BASE_URL)
+
+    selector = page.get_by_text('Running...')
+    selector.wait_for(state='detached')
+
+    expect(page).to_have_title("Dianna's dashboard")
+    for selector in (
+            page.get_by_role('img', name='0'),
+            page.get_by_text('More information'),
+    ):
+        expect(selector).to_be_visible()
+
